STY: moved code for DialogViewMetadata to a separate file
